Exporting misc controllers as class
